command-line interface
