+++ setup.py
@@ -23,5 +23,9 @@ setup(
     install_requires=[
         'requests',
         'sslyze @ git+https://github.com/fabian-hk/sslyze.git@tls_profiler#egg=sslyze',
+        'cryptography>=2.8',
     ],
+    entry_points={
+        "console_scripts": ["tlsprofiler=tlsprofiler.cli:main"]
+    },
 )
